# services/mapi-xbt/docs/indexer/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bitcoin - Blockchain Indexer API",
    "description": "This API provides core indexer endpoints with support for Bitcoin metaprotocols by delivering real-time, rollback-protected access to Bitcoin's UTXO data, enabling developers to build responsive and reliable blockchain applications without managing complex infrastructure.\n\n#### Key Features:\n- **Real-Time Data with Rollback Protection:** Ensures data accuracy by handling chain reorganizations gracefully, providing live data without sacrificing integrity.\n- **Comprehensive UTXO Indexing:** Specialized pipelines extract, match, and process on-chain information, including handling rollbacks, to provide accurate and up-to-date data.\n\n#### Key Benefits for Developers:\nBy abstracting the complexities of blockchain data retrieval and processing, the Bitcoin Indexer API empowers developers to focus on building innovative applications with confidence in fast and reliable access to historical chain data.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.txt"
    },
    "version": "v0.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local instance"
    }
  ],
  "paths": {
...
    "/assets/brc20/{ticker}": {
      "get": {
        "tags": [
          "BRC20"
        ],
        "summary": "BRC20 Info",
        "description": "Information about a BRC20 token’s metadata and current state, including its symbol, deployment details, minting rules, total holders, and minted supply.",
        "operationId": "brc20_info",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "description": "BRC20 ticker string",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "FCTB"
          }
        ],
        "responses": {
          "200": {
            "description": "Requested data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimestampedBrc20Info"
                },
                "example": {
                  "data": {
                    "deploy_inscription": "3d983a724310f511cfca0031ee2b980b474a0abe1e7e995b7e6d2873e2cbfd5fi0",
                    "holders": 7,
                    "minted_supply": "19420000.000000000000000000",
                    "terms": {
                      "dec": 18,
                      "limit": "1000000.000000000000000000",
                      "max": "21000000.000000000000000000",
                      "self_mint": false
                    },
                    "ticker": "fctb",
                    "ticker_hex": "66637462"
                  },
                  "last_updated": {
                    "block_hash": "0000000000000000000214bfa0a73e1cc7663917a933bfb1c66a6613f88dabdd",
                    "block_height": 851556
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameters"
          },
          "404": {
            "description": "Requested entity not found on-chain"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/assets/collections/{collection_symbol}/metadata": {
      "get": {
        "tags": [
          "Inscriptions"
        ],
        "summary": "Collection Metadata by Collection Symbol",
        "description": "Provides metadata for a given inscription collection symbol, including its name, image, supply, and external links. This is useful for rendering collection summaries or for display in marketplaces and aggregators.",
        "operationId": "collection_metadata_by_collection_symbol",
        "parameters": [
          {
            "name": "collection_symbol",
            "in": "path",
            "description": "Collection symbol (UTF-8)",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "twick"
          }
        ],
        "responses": {
          "200": {
            "description": "Requested data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimestampedCollectionMetadataBySymbol"
                },
                "example": {
                  "data": {
                    "chain": "btc",
                    "createdAt": "Sun, 17 Sep 2023 06:38:15 GMT",
                    "description": "A collection of 1/1 art inscribed forever on the Bitcoin Blockchain. All 1/1s are inscribed on special sats.",
                    "discordLink": "",
                    "imageURI": "https://creator-hub-prod.s3.us-east-2.amazonaws.com/ord-twick_pfp_1705357607896.jpeg",
                    "inscriptionIcon": "736d7692fb2b405f7efdd05b57787a10bb0385b7dd394aa67654d62119e9d825i0",
                    "labels": [],
                    "max_inscription_number": "63111093",
                    "min_inscription_number": "31921195",
                    "name": "1/1 ART BY TWICK",
                    "supply": 10,
                    "symbol": "twick",
                    "twitterLink": "https://twitter.com/Twickert_",
                    "websiteLink": "https://www.artbytwick.com/"
                  },
                  "last_updated": {
                    "block_hash": "00000000000000000001998e2059bcbb25f76fd0ef39db8ddfc5c31c5ea95f1f",
                    "block_height": 876644
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameters"
          },
          "404": {
            "description": "Requested entity not found on-chain"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Brc20Info": {
        "type": "object",
        "required": [
          "ticker",
          "ticker_hex",
          "deploy_inscription",
          "holders",
          "minted_supply",
          "terms"
        ],
        "properties": {
          "deploy_inscription": {
            "type": "string"
          },
          "holders": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "minted_supply": {
            "type": "string"
          },
          "terms": {
            "$ref": "#/components/schemas/Brc20Terms"
          },
          "ticker": {
            "type": "string"
          },
          "ticker_hex": {
            "type": "string"
          }
        }
      },
      "Brc20Terms": {
        "type": "object",
        "required": [
          "max",
          "limit",
          "dec",
          "self_mint"
        ],
        "properties": {
          "dec": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "limit": {
            "type": "string"
          },
          "max": {
            "type": "string"
          },
          "self_mint": {
            "type": "boolean"
          }
        }
      },
...
      "ChainTip": {
        "type": "object",
        "required": [
          "block_hash",
          "block_height"
        ],
        "properties": {
          "block_hash": {
            "type": "string",
            "description": "The hash of the block",
            "example": "0000000000000000000a7f3b7b6b6e1d9a18db65a3b4a3f4f3bcb2e1f1b2d3e7"
          },
          "block_height": {
            "type": "integer",
            "format": "int64",
            "description": "The height of the block in the blockchain",
            "example": 707000,
            "minimum": 0
          }
        }
      },
      "CollectionMetadata": {
        "type": "object",
        "required": [
          "symbol",
          "name",
          "imageURI",
          "chain",
          "inscriptionIcon",
          "description",
          "supply",
          "twitterLink",
          "discordLink",
          "websiteLink",
          "min_inscription_number",
          "max_inscription_number",
          "createdAt"
        ],
        "properties": {
          "chain": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "discordLink": {
            "type": "string"
          },
          "imageURI": {
            "type": "string"
          },
          "inscriptionIcon": {
            "type": "string"
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "max_inscription_number": {
            "type": "string"
          },
          "min_inscription_number": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "supply": {},
          "symbol": {
            "type": "string"
          },
          "twitterLink": {
            "type": "string"
          },
          "websiteLink": {
            "type": "string"
          }
        }
      },
...
      "TimestampedBrc20Info": {
        "type": "object",
        "required": [
          "data",
          "last_updated"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Brc20Info"
          },
          "last_updated": {
            "$ref": "#/components/schemas/ChainTip"
          }
        }
      },
...
      "TimestampedCollectionMetadataBySymbol": {
        "type": "object",
        "required": [
          "data",
          "last_updated"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CollectionMetadata"
          },
          "last_updated": {
            "$ref": "#/components/schemas/ChainTip"
          }
        }
      },
...
    }
  }
}